fix remaining flake8 things, ignore the rest
