+++ textattack/commands/__init__.py
@@ -1,17 +1,3 @@
 from abc import ABC, abstractmethod
-
-# from argparse import ArgumentParser, HelpFormatter
-
-
-class TextAttackCommand(ABC):
-    @staticmethod
-    @abstractmethod
-    def register_subcommand(parser):
-        raise NotImplementedError()
-
-    @abstractmethod
-    def run(self):
-        raise NotImplementedError()
-
-
+from .textattack_command import TextAttackCommand
 from . import textattack_cli
