+++ textattack/commands/attack/attack_args.py
@@ -156,7 +156,7 @@ HUGGINGFACE_DATASET_BY_MODEL = {
     ),
     "roberta-base-imdb": ("textattack/roberta-base-imdb", ("imdb", None, "test"),),
     "roberta-base-mr": (
-        "textattack/textattack/roberta-base-rotten-tomatoes",
+        "textattack/roberta-base-rotten-tomatoes",
         ("rotten_tomatoes", None, "test"),
     ),
     "roberta-base-mrpc": (
@@ -179,10 +179,6 @@ HUGGINGFACE_DATASET_BY_MODEL = {
     "roberta-base-wnli": (
         "textattack/roberta-base-WNLI",
         ("glue", "wnli", "validation"),
-    ),
-    "roberta-base-mr": (
-        "textattack/roberta-base-rotten-tomatoes",
-        ("rotten_tomatoes", None, "test"),
     ),
     #
     # albert-base-v2 (ALBERT is cased by default)
