+++ textattack/commands/attack/run_attack_single_threaded.py
@@ -146,7 +146,3 @@ def run(args, checkpoint=None):
         textattack.shared.logger.info(f"Attack time: {time.time() - load_time}s")
 
         return attack_log_manager.results
-
-
-if __name__ == "__main__":
-    run(get_args())
